% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrada_scopus.R
\name{entrada_scopus}
\alias{entrada_scopus}
\title{Function that generates the test database}
\usage{
entrada_scopus(path_date, cabec, quoaspa)
}
\arguments{
\item{path_date}{is a directory path containing the csv file}

\item{cabec}{is the header for files in bib format}

\item{quoaspa}{is the separator for files in bib format}
}
\value{
This function return is a dataframe with database
}
\description{
The example_database function reads a csv file available on the computer system and returns a dataframe as an example of a database to be used by the functions of the bibliorefer package
}
\examples{

#Call the example_database function

file_db <- system.file("extdata","scopus.bib", package = "bibliorefer")
cabec = TRUE
quoaspa = ""
date_sreference <- entrada_scopus(file_db, cabec, quoaspa)
date_sreference

}
\references{
Aria, M. & Cuccurullo, C. (2017) bibliometrix: An R-tool for comprehensive science mapping analysis, Journal of Informetrics, 11(4), pp 959-975, Elsevier
}
