% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrada_wos.R
\name{entrada_wos}
\alias{entrada_wos}
\title{Function that generates the test database}
\usage{
entrada_wos(path_date, cabec_wos, quoaspa_wos)
}
\arguments{
\item{path_date}{is a directory path containing the bib file}

\item{cabec_wos}{is the header for wos files in bib format}

\item{quoaspa_wos}{is the separator wos for files in bib format}
}
\value{
This function return is a dataframe with database
}
\description{
The example_database function reads a csv file available on the computer system and returns a dataframe as an example of a database to be used by the functions of the bibliorefer package
}
\examples{

#Call the example_database function

file_db <- system.file("extdata","savedrecs.bib", package = "bibliorefer")
cabec_wos = TRUE
quoaspa_wos = ""
date_wosrefer <- entrada_wos(file_db, cabec_wos, quoaspa_wos)
date_wosrefer

}
\references{
Aria, M. & Cuccurullo, C. (2017) bibliometrix: An R-tool for comprehensive science mapping analysis, Journal of Informetrics, 11(4), pp 959-975, Elsevier
}
