% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produc_cientific.R
\name{produc_cientific}
\alias{produc_cientific}
\title{Function of annual scientific production}
\usage{
produc_cientific(input_date)
}
\arguments{
\item{input_date}{is a dataframe with the scientific production database}
}
\value{
The function return a dataframe with the dates and the number of articles published in each of them
}
\description{
This function shows the evolution of scientific production. The function receives a dataframe with scientific production data and returns a dataframe with the dates and number of articles published on each date
}
\examples{

#Call the produc_cientific function
file_db <- system.file("extdata","example_database.csv", package = "bibliorefer")
separator <- ","
input_date <- example_database(file_db, separator)
prod_ct <- produc_cientific(input_date)
prod_ct


}
\references{
Evolution of number of citations per article in Materials Science: possible causes and effect on the impact factor of journals. Ariza-Guerrero, A.M, Blázquez, J.S. Scientometrics, 128(12), pp. 6589–6609. 2023.
}
