% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabel_gera.R
\name{tabel_gera}
\alias{tabel_gera}
\title{Function of scientific information}
\usage{
tabel_gera(input_date)
}
\arguments{
\item{input_date}{is a dataframe with the scientific production database}
}
\value{
The function return a list object containing three dataframes with ordered lists of articles, authors and journals.
}
\description{
This function receives a dataframe with scientific production data and returns a list object with information about articles, authors and journals.
}
\examples{

#Call the tabel_gera function
file_db <- system.file("extdata","example_database.csv",package = "bibliorefer")
separator <- ","
input_date <- example_database(file_db, separator)
scientific_inf <- tabel_gera(input_date)
scientific_inf


}
\references{
Aria, M. & Cuccurullo, C. (2017) bibliometrix: An R-tool for comprehensive science mapping analysis, Journal of Informetrics, 11(4), pp 959-975, Elsevier
}
